+++ python/semantic_kernel/connectors/ai/ai_exception.py
