+++ python/semantic_kernel/connectors/ai/embeddings/embedding_generator_base.py
