+++ python/semantic_kernel/connectors/ai/embeddings/embedding_index_base.py
