+++ python/semantic_kernel/connectors/ai/open_ai/__init__.py
@@ -1,0 +1,27 @@
+# Copyright (c) Microsoft. All rights reserved.
+
+from semantic_kernel.connectors.ai.open_ai.services.azure_chat_completion import (
+    AzureChatCompletion,
+)
+from semantic_kernel.connectors.ai.open_ai.services.azure_text_completion import (
+    AzureTextCompletion,
+)
+from semantic_kernel.connectors.ai.open_ai.services.azure_text_embedding import AzureTextEmbedding
+from semantic_kernel.connectors.ai.open_ai.services.open_ai_chat_completion import (
+    OpenAIChatCompletion,
+)
+from semantic_kernel.connectors.ai.open_ai.services.open_ai_text_completion import (
+    OpenAITextCompletion,
+)
+from semantic_kernel.connectors.ai.open_ai.services.open_ai_text_embedding import (
+    OpenAITextEmbedding,
+)
+
+__all__ = [
+    "OpenAITextCompletion",
+    "OpenAIChatCompletion",
+    "OpenAITextEmbedding",
+    "AzureTextCompletion",
+    "AzureChatCompletion",
+    "AzureTextEmbedding",
+]
