+++ python/semantic_kernel/connectors/ai/open_ai/services/open_ai_chat_completion.py
@@ -3,9 +3,9 @@
 from logging import Logger
 from typing import Any, List, Optional, Tuple
 
-from semantic_kernel.ai.ai_exception import AIException
-from semantic_kernel.ai.chat_completion_client_base import ChatCompletionClientBase
-from semantic_kernel.ai.chat_request_settings import ChatRequestSettings
+from semantic_kernel.connectors.ai.ai_exception import AIException
+from semantic_kernel.connectors.ai.chat_completion_client_base import ChatCompletionClientBase
+from semantic_kernel.connectors.ai.chat_request_settings import ChatRequestSettings
 from semantic_kernel.utils.null_logger import NullLogger
