+++ python/semantic_kernel/connectors/ai/open_ai/services/open_ai_text_completion.py
@@ -3,9 +3,9 @@
 from logging import Logger
 from typing import Any, Optional
 
-from semantic_kernel.ai.ai_exception import AIException
-from semantic_kernel.ai.complete_request_settings import CompleteRequestSettings
-from semantic_kernel.ai.text_completion_client_base import TextCompletionClientBase
+from semantic_kernel.connectors.ai.ai_exception import AIException
+from semantic_kernel.connectors.ai.complete_request_settings import CompleteRequestSettings
+from semantic_kernel.connectors.ai.text_completion_client_base import TextCompletionClientBase
 from semantic_kernel.utils.null_logger import NullLogger
