+++ python/semantic_kernel/connectors/ai/text_completion_client_base.py
@@ -5,7 +5,7 @@ from logging import Logger
 from typing import TYPE_CHECKING
 
 if TYPE_CHECKING:
-    from semantic_kernel.ai.complete_request_settings import CompleteRequestSettings
+    from semantic_kernel.connectors.ai.complete_request_settings import CompleteRequestSettings
 
 
 class TextCompletionClientBase(ABC):
