+++ python/semantic_kernel/memory/memory_store_base.py
@@ -2,7 +2,7 @@
 
 from abc import ABC
 
-from semantic_kernel.ai.embeddings.embedding_index_base import EmbeddingIndexBase
+from semantic_kernel.connectors.ai.embeddings.embedding_index_base import EmbeddingIndexBase
 from semantic_kernel.memory.storage.data_store_base import DataStoreBase
