+++ python/tests/end-to-end/skills_from_dir.py
@@ -4,7 +4,7 @@ import asyncio
 import os
 
 import semantic_kernel as sk
-import semantic_kernel.ai.open_ai as sk_oai
+import semantic_kernel.connectors.ai.open_ai as sk_oai
 
 kernel = sk.Kernel()
